refactor: useData 훅을 DataContext에 작성
